+++ HISTORY.rst
@@ -2,7 +2,17 @@
 History
 =======
 
-0.1.0 (2019-10-01)
+0.8.2 (2020-03-12)
+------------------
+
+* Fixed tests, general bugfixes
+
+0.8.1 (2020-03-11)
+------------------
+
+* Improved service files.
+
+0.8.0 (2020-03-09)
 ------------------
 
 * First release on PyPI.